# packages/server-api/openapi.yaml
openapi: 3.1.0
info:
  title: Awesome Omni Skills Catalog API
  version: 0.12.9
  description: Read-only API for skill discovery, manifests, bundles, archives, and install planning.
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Taxonomy:
      type: object
      properties:
        raw_category:
          type: string
        canonical_category:
          type: string
        inferred_category:
          type: string
        category_source:
          type: string
    ListedSkill:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        display_name: { type: string }
        description: { type: string }
        version: { type: string }
        category: { type: string }
        raw_category: { type: string }
        canonical_category: { type: string }
        taxonomy:
          $ref: '#/components/schemas/Taxonomy'
        tags:
          type: array
          items: { type: string }
        complexity: { type: string }
        risk: { type: string }
        tools:
          type: array
          items: { type: string }
        entrypoint: { type: string }
        manifest_path: { type: string }
        metadata_path: { type: string }
        skill_level: { type: integer }
        skill_level_label: { type: string }
        has_scripts: { type: boolean }
        has_extra_files: { type: boolean }
        best_practices_score: { type: integer }
        best_practices_tier: { type: string }
        quality_score: { type: integer }
        quality_tier: { type: string }
        security_score: { type: integer }
        security_tier: { type: string }
        security_status: { type: string }
        archives_count: { type: integer }
        validation_status: { type: string }
        date_updated:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        generated_at:
          type: string
          format: date-time
    SkillArtifact:
      type: object
      properties:
        path: { type: string }
        size_bytes: { type: integer }
        download_url:
          type: [string, "null"]
    ArchiveSignature:
      type: object
      properties:
        path: { type: string }
        size_bytes: { type: integer }
        download_url:
          type: [string, "null"]
    SkillArchive:
      type: object
      properties:
        format: { type: string }
        file_name: { type: string }
        path: { type: string }
        size_bytes: { type: integer }
        download_url:
          type: [string, "null"]
        signature:
          oneOf:
            - $ref: '#/components/schemas/ArchiveSignature'
            - type: "null"
    ArchiveChecksums:
      type: object
      properties:
        path: { type: string }
        file_name: { type: string }
        size_bytes: { type: integer }
    SkillMaturity:
      type: object
      properties:
        skill_level: { type: integer }
        skill_level_label: { type: string }
    ClassificationScore:
      type: object
      properties:
        score: { type: integer }
    SecurityClassification:
      type: object
      properties:
        score: { type: integer }
        status: { type: string }
    SkillClassification:
      type: object
      properties:
        maturity:
          $ref: '#/components/schemas/SkillMaturity'
        best_practices:
          $ref: '#/components/schemas/ClassificationScore'
        quality:
          $ref: '#/components/schemas/ClassificationScore'
        security:
          $ref: '#/components/schemas/SecurityClassification'
    SkillInstallTarget:
      type: object
      properties:
        tool: { type: string }
        path: { type: string }
    SkillCompatibility:
      type: object
      properties:
        tools:
          type: array
          items: { type: string }
        install_targets:
          type: array
          items:
            $ref: '#/components/schemas/SkillInstallTarget'
    InstallRecipe:
      type: object
      properties:
        tool: { type: string }
        command: { type: string }
        scope: { type: string }
        default_path: { type: string }
        behavior: { type: string }
    SkillInstallDefinition:
      type: object
      properties:
        recipes:
          type: array
          items:
            $ref: '#/components/schemas/InstallRecipe'
    ManifestPaths:
      type: object
      properties:
        manifest: { type: string }
    SkillLinks:
      type: object
      properties:
        self:
          type: [string, "null"]
        downloads:
          type: [string, "null"]
        manifest:
          type: [string, "null"]
        entrypoint:
          type: [string, "null"]
        artifacts:
          type: [string, "null"]
        archives:
          type: [string, "null"]
        archive_checksums:
          type: [string, "null"]
    ChecksumsMap:
      type: object
      additionalProperties:
        type: string
    ManifestSkill:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        display_name: { type: string }
        description: { type: string }
        category: { type: string }
        raw_category: { type: string }
        taxonomy:
          $ref: '#/components/schemas/Taxonomy'
        tags:
          type: array
          items: { type: string }
        risk: { type: string }
        entrypoint: { type: string }
        paths:
          $ref: '#/components/schemas/ManifestPaths'
        artifacts:
          type: array
          items:
            $ref: '#/components/schemas/SkillArtifact'
        archives:
          type: array
          items:
            $ref: '#/components/schemas/SkillArchive'
        archive_checksums:
          oneOf:
            - $ref: '#/components/schemas/ArchiveChecksums'
            - type: "null"
        checksums:
          $ref: '#/components/schemas/ChecksumsMap'
        compatibility:
          $ref: '#/components/schemas/SkillCompatibility'
        install:
          $ref: '#/components/schemas/SkillInstallDefinition'
        classification:
          $ref: '#/components/schemas/SkillClassification'
        links:
          oneOf:
            - $ref: '#/components/schemas/SkillLinks'
            - type: "null"
...
    SkillListFilters:
      type: object
      properties:
        category:
          type: [string, "null"]
        tool:
          type: [string, "null"]
        risk:
          type: [string, "null"]
        min_quality_score:
          type: [integer, "null"]
        min_best_practices_score:
          type: [integer, "null"]
        min_skill_level:
          type: [integer, "null"]
        min_security_score:
          type: [integer, "null"]
        validation_status:
          type: [string, "null"]
        security_status:
          type: [string, "null"]
    SkillListResponse:
      type: object
      properties:
        total: { type: integer }
        offset: { type: integer }
        limit: { type: integer }
        sort: { type: string }
        order: { type: string, enum: [asc, desc] }
        filters:
          $ref: '#/components/schemas/SkillListFilters'
        results:
          type: array
          items:
            $ref: '#/components/schemas/ListedSkill'
    SkillArtifactsResponse:
      type: object
      properties:
        id: { type: string }
        links:
          $ref: '#/components/schemas/SkillLinks'
        checksums:
          $ref: '#/components/schemas/ChecksumsMap'
        artifacts:
          type: array
          items:
            $ref: '#/components/schemas/SkillArtifact'
    SkillArchivesResponse:
      type: object
      properties:
        id: { type: string }
        links:
          $ref: '#/components/schemas/SkillLinks'
        archive_checksums:
          oneOf:
            - $ref: '#/components/schemas/ArchiveChecksums'
            - type: "null"
        archives:
          type: array
          items:
            $ref: '#/components/schemas/SkillArchive'
    SkillDownloadsResponse:
      type: object
      properties:
        id: { type: string }
        links:
          $ref: '#/components/schemas/SkillLinks'
        checksums:
          $ref: '#/components/schemas/ChecksumsMap'
        artifacts:
          type: array
          items:
            $ref: '#/components/schemas/SkillArtifact'
        archives:
          type: array
          items:
            $ref: '#/components/schemas/SkillArchive'
        archive_checksums:
          oneOf:
            - $ref: '#/components/schemas/ArchiveChecksums'
            - type: "null"
    CompareSkillSummary:
      type: object
      properties:
        id: { type: string }
        display_name: { type: string }
        description: { type: string }
        category: { type: string }
        tools:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
        risk: { type: string }
        manifest_path: { type: string }
    CompareShared:
      type: object
      properties:
        categories:
          type: array
          items: { type: string }
        tools:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
    CompareDifference:
      type: object
      properties:
        id: { type: string }
        unique_tools:
          type: array
          items: { type: string }
        unique_tags:
          type: array
          items: { type: string }
    CompareResponse:
      type: object
      properties:
        skills:
          type: array
          items:
            $ref: '#/components/schemas/CompareSkillSummary'
        shared:
          $ref: '#/components/schemas/CompareShared'
        differences:
          type: array
          items:
            $ref: '#/components/schemas/CompareDifference'
...
    InstallPlanRequest:
      type: object
      properties:
        skill_ids:
          type: array
          items: { type: string }
        bundle_ids:
          type: array
          items: { type: string }
        tools:
          type: array
          items: { type: string }
        target_path:
          type: [string, "null"]
        dry_run:
          type: boolean
    InstallStep:
      type: object
      properties:
        action: { type: string }
        command: { type: string }
        description: { type: string }
    InstallSelectionBundle:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        available_skill_ids:
          type: array
          items: { type: string }
        missing_skill_ids:
          type: array
          items: { type: string }
    InstallSelectionSkill:
      type: object
      properties:
        id: { type: string }
        display_name: { type: string }
        manifest_path: { type: string }
        checksums:
          $ref: '#/components/schemas/ChecksumsMap'
        links:
          $ref: '#/components/schemas/SkillLinks'
        archives:
          type: array
          items:
            $ref: '#/components/schemas/SkillArchive'
        classification:
          $ref: '#/components/schemas/SkillClassification'
    ClientRecipe:
      type: object
      properties:
        tool: { type: string }
        command: { type: string }
        scope: { type: string }
        default_path: { type: string }
        behavior: { type: string }
    InstallDownloads:
      type: object
      properties:
        catalog:
          type: [string, "null"]
        artifacts_available:
          type: boolean
    InstallPlanResponse:
      type: object
      properties:
        dry_run: { type: boolean }
        selected_bundles:
          type: array
          items:
            $ref: '#/components/schemas/InstallSelectionBundle'
        selected_skills:
          type: array
          items:
            $ref: '#/components/schemas/InstallSelectionSkill'
        install_scope: { type: string }
        target_path:
          type: [string, "null"]
        target_clients:
          type: array
          items: { type: string }
        commands:
          type: array
          items: { type: string }
        recipes:
          type: array
          items:
            $ref: '#/components/schemas/ClientRecipe'
        warnings:
          type: array
          items: { type: string }
        downloads:
          $ref: '#/components/schemas/InstallDownloads'
        next_steps:
          type: array
          items:
            $ref: '#/components/schemas/InstallStep'
...
paths:
...
  /v1/skills:
    get:
      summary: List published skills with filters, sorting, and pagination
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: tool
          in: query
          schema: { type: string }
        - name: risk
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string }
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - name: limit
          in: query
          schema: { type: integer }
        - name: offset
          in: query
          schema: { type: integer }
        - name: min_quality
          in: query
          schema: { type: integer }
        - name: min_best_practices
          in: query
          schema: { type: integer }
        - name: min_level
          in: query
          schema: { type: integer }
        - name: min_security
          in: query
          schema: { type: integer }
        - name: validation_status
          in: query
          schema: { type: string }
        - name: security_status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Filtered skill list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillListResponse'
  /v1/skills/{id}:
    get:
      summary: Resolve a manifest-backed skill definition
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Manifest-backed skill payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManifestSkill'
        "404":
          description: Skill not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/skills/{id}/artifacts:
    get:
      summary: List artifact files for a skill
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Artifact list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillArtifactsResponse'
        "404":
          description: Skill not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/skills/{id}/archives:
    get:
      summary: List downloadable archives for a skill
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Archive list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillArchivesResponse'
        "404":
          description: Skill not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/skills/{id}/downloads:
    get:
      summary: Resolve full download metadata for a skill
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Download metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillDownloadsResponse'
        "404":
          description: Skill not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/skills/{id}/download/manifest:
    get:
      summary: Download a skill manifest JSON file
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Manifest file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManifestSkill'
        "404":
          description: Skill not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/skills/{id}/download/entrypoint:
    get:
      summary: Download the SKILL.md entrypoint for a skill
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Markdown file
          content:
            text/markdown:
              schema:
                type: string
        "404":
          description: Skill not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/skills/{id}/download/artifact:
    get:
      summary: Download a specific artifact from a skill
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Artifact file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Artifact not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/skills/{id}/download/archive:
    get:
      summary: Download a packaged archive for a skill
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: format
          in: query
          required: false
          schema:
            type: string
            default: zip
      responses:
        "200":
          description: Archive file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Archive not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/skills/{id}/download/archive/signature:
    get:
      summary: Download the signature file for a skill archive
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: format
          in: query
          required: false
          schema:
            type: string
            default: zip
      responses:
        "200":
          description: Signature file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Signature not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/skills/{id}/download/archive/checksums:
    get:
      summary: Download the archive checksum manifest for a skill
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Checksum manifest
          content:
            text/plain:
              schema:
                type: string
        "404":
          description: Checksums not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/families/{id}/variants:
    get:
      summary: List the variants of a skill family
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Variant list for the family
          content:
            application/json:
              schema:
                type: object
                properties:
                  family_id:
                    type: string
                  total:
                    type: integer
                  results:
                    type: array
                    items:
                      type: object
        "404":
          description: Family not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/search:
    get:
      summary: Search skills using the same filters and relevance engine as `/v1/skills`
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: tool
          in: query
          schema: { type: string }
        - name: risk
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string }
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - name: limit
          in: query
          schema: { type: integer }
        - name: offset
          in: query
          schema: { type: integer }
        - name: min_quality
          in: query
          schema: { type: integer }
        - name: min_best_practices
          in: query
          schema: { type: integer }
        - name: min_level
          in: query
          schema: { type: integer }
        - name: min_security
          in: query
          schema: { type: integer }
        - name: validation_status
          in: query
          schema: { type: string }
        - name: security_status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillListResponse'
  /v1/compare:
    get:
      summary: Compare multiple skills by comma-separated IDs
      parameters:
        - name: ids
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Comparison payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompareResponse'
...
  /v1/install/plan:
    post:
      summary: Build an install plan for selected skills, bundles, targets, or custom paths
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstallPlanRequest'
      responses:
        "200":
          description: Install plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallPlanResponse'